input validation
added input validation to the game. The game will only accept your turn if you enter a number between 1 and 3 and if the amount you are trying to take is actually available. Our professor wants error catching so we have to accept input in the form of a string and then parse it into an integer.