Added showEnding()
The game will now end by displaying a winner, a loser, and the total of each. It can also handle a tie. If you look at the code from showTotals() you'll notice that the logic at the start is actually identical to showEnding(), but showEnding() has been written in a much more compact manner.