enemy choice
Using the technique we learned about in class on wednesday, each of the piles is passed as an argument to the "enemy player" and evaluated. The enemy player prefers A, B and C in order, so no need for complex validation and do...while loops. We try to validate each pile once, and once we find a valid pile we're finished.