AI upgrade
The AI will try to force the player to take the last value when it wants the jackpot and force the player to leave at least 1 when it does not want the jackpot(allowing the AI to avoid taking it). The pile it chooses is now random, which is certainly possible to optimize but its not time to submit Part 2 yet so I'm not coding it. Feel free to do that if you like.